upload excel sheet and download in animal

+++ src/Components/Mating/MatingTable.jsx
@@ -17,7 +17,7 @@ function MatingTable() {
     const [currentPage, setCurrentPage] = useState(1); 
     const [totalPages, setTotalPages] = useState(0);
     const animalsPerPage = 10; 
-    const [searchCriteria, setSearchCriteria] = useState({ tagId: '', matingDate: '', sonarRsult: '' });
+    const [searchCriteria, setSearchCriteria] = useState({ tagId: '', matingDate: '', sonarRsult: '', animalType: '' });
     const [matings, setMating] = useState([]);
 
     async function fetchMating() {
@@ -27,6 +27,7 @@ function MatingTable() {
                 tagId: searchCriteria.tagId,
                 matingDate: searchCriteria.matingDate,
                 sonarRsult: searchCriteria.sonarRsult,
+                animalType: searchCriteria.animalType,
             };
             const { data } = await getMating(currentPage, animalsPerPage, filters);
             setMating(data.data.mating);
@@ -40,31 +41,26 @@ function MatingTable() {
 
     useEffect(() => {
         fetchMating();
-    }, [currentPage]); // Fetch mating data on page change
+    }, [currentPage]); 
 
     const deleteItem = async (id) => {
         try {
             await deleteMating(id);
             setMating((prevMating) => prevMating.filter((mating) => mating._id !== id));
         } catch (error) {
-            console.error("Failed to delete mating:", error);
             Swal.fire('Error', 'Failed to delete mating', 'error');
         }
     };
 
-    const handleClick = (id) => {
+    const confirmDelete = (id) => {
         Swal.fire({
             title: "هل تريد الاستمرار؟",
             icon: "question",
-            iconHtml: "؟",
             confirmButtonText: "نعم",
             cancelButtonText: "لا",
             showCancelButton: true,
-            showCloseButton: true
         }).then((result) => {
-            if (result.isConfirmed) {
-                deleteItem(id);
-            }
+            if (result.isConfirmed) deleteItem(id);
         });
     };
 
@@ -73,14 +69,11 @@ function MatingTable() {
     };
 
     const handleSearch = () => {
-        setCurrentPage(1); // Reset to first page when searching
-        fetchMating(); // Fetch mating data with the current search criteria
+        setCurrentPage(1); 
+        fetchMating(); 
     };
 
-    // Pagination function
-    const paginate = (pageNumber) => {
-        setCurrentPage(pageNumber);
-    };
+    const paginate = (pageNumber) => setCurrentPage(pageNumber);
 
     return (
         <>
@@ -91,68 +84,86 @@ function MatingTable() {
             ) : (
                 <div className="container">
                     <div className="title2">Mating</div>
-                    <div className="d-flex justify-content-between align-items-center mt-4 flex-wrap">
-                        <div className="d-flex flex-wrap" style={{ flex: 1 }}>
-                            <input
-                                type="text"
-                                className="form-control me-2 mb-2"
-                                placeholder="Search by Tag ID"
-                                value={searchCriteria.tagId}
-                                onChange={(e) => setSearchCriteria({ ...searchCriteria, tagId: e.target.value })}
-                                style={{ flex: '1' }}
-                                aria-label="Search by Tag ID"
-                            />
-                            <input
-                                type="date"
-                                className="form-control me-2 mb-2"
-                                value={searchCriteria.matingDate}
-                                onChange={(e) => setSearchCriteria({ ...searchCriteria, matingDate: e.target.value })}
-                                style={{ flex: '1' }}
-                                aria-label="Search by Mating Date"
-                            />
-                            <input
-                                type="text"
-                                className="form-control me-2 mb-2"
-                                placeholder="Search by Sonar Result"
-                                value={searchCriteria.sonarRsult}
-                                onChange={(e) => setSearchCriteria({ ...searchCriteria, sonarRsult: e.target.value })}
-                                style={{ flex: '1' }}
-                                aria-label="Search by Sonar Result"
-                            />
-                            <button 
-                                className="btn mb-2 me-2" 
-                                onClick={handleSearch}
-                                style={{ backgroundColor: '#81a9d1', borderColor: '#81a9d1', color: 'white' }}
-                            >
-                                <i className="fas fa-search"></i>
-                            </button>
-                        </div>
-                        <Link to='/mating'>
-                            <button type="button" className="btn btn-secondary btn-lg active button2 mb-2">
+                    
+                    <Link to='/mating'>
+                            <button type="button" className="btn btn-secondary button2 btn-lg active mt-3">
                                 <MdOutlineAddToPhotos /> Add New Mating
                             </button>
                         </Link>
+                    <div className="d-flex justify-content-between align-items-center mt-4 flex-wrap">
+            <div className="d-flex align-items-center" style={{ flex: 1, flexWrap: 'nowrap', gap: '1rem' }}>
+    <input
+        type="text"
+        className="form-control"
+        style={{ minWidth: '150px' }}
+        placeholder="Search by Tag ID"
+        value={searchCriteria.tagId}
+        onChange={(e) => setSearchCriteria({ ...searchCriteria, tagId: e.target.value })}
+        aria-label="Search by Tag ID"
+    />
+    
+    <select
+        value={searchCriteria.animalType}
+        className="form-select"
+        style={{ minWidth: '150px' }}
+        onChange={(e) => setSearchCriteria({ ...searchCriteria, animalType: e.target.value })}
+        aria-label="Filter by Animal Type"
+    >
+        <option value="">Animal Type</option>
+        <option value="goat">Goat</option>
+        <option value="sheep">Sheep</option>
+    </select>
+
+    <input
+        type="date"
+        className="form-control"
+        style={{ minWidth: '150px' }}
+        placeholder="Mating Date"
+        value={searchCriteria.matingDate}
+        onChange={(e) => setSearchCriteria({ ...searchCriteria, matingDate: e.target.value })}
+        aria-label="Search by Mating Date"
+    />
+
+    <input
+        type="text"
+        className="form-control"
+        style={{ minWidth: '150px' }}
+        placeholder="Search by Sonar Result"
+        value={searchCriteria.sonarRsult}
+        onChange={(e) => setSearchCriteria({ ...searchCriteria, sonarRsult: e.target.value })}
+        aria-label="Search by Sonar Result"
+    />
+    
+    <button 
+        className="btn me-3" 
+        onClick={handleSearch}
+        style={{ backgroundColor: '#81a9d1', color: 'white' }}
+    >
+        <i className="fas fa-search"></i>
+    </button>
+</div>
+
                     </div>
                     
-                    <table className="table table-striped mt-6">
+                    <table className="table table-striped mt-4">
                         <thead>
                             <tr>
-                                <th scope="col">#</th>
-                                <th scope="col">Tag ID</th>
-                                <th scope="col">Male Tag Id</th>
-                                <th scope="col">Mating Type</th>
-                                <th scope="col">Mating Date</th>
-                                <th scope="col">Sonar Date</th>
-                                <th scope="col">Sonar Result</th>
-                                <th scope="col">Expected Delivery Date</th>
-                                <th scope="col">Edit Mating</th>
-                                <th scope="col">Remove Mating</th>
+                                <th>#</th>
+                                <th>Tag ID</th>
+                                <th>Male Tag Id</th>
+                                <th>Mating Type</th>
+                                <th>Mating Date</th>
+                                <th>Sonar Date</th>
+                                <th>Sonar Result</th>
+                                <th>Expected Delivery Date</th>
+                                <th>Edit Mating</th>
+                                <th>Remove Mating</th>
                             </tr>
                         </thead>
                         <tbody>
                             {matings.map((mating, index) => (
-                                <tr key={`${mating._id}-${index}`}>
-                                    <th scope="row">{(currentPage - 1) * animalsPerPage + index + 1}</th>
+                                <tr key={mating._id}>
+                                    <td>{(currentPage - 1) * animalsPerPage + index + 1}</td>
                                     <td>{mating.tagId}</td>
                                     <td>{mating.maleTag_id}</td>
                                     <td>{mating.matingType}</td>
@@ -160,10 +171,11 @@ function MatingTable() {
                                     <td>{mating.sonarDate ? mating.sonarDate.split('T')[0] : NO_DATE}</td>
                                     <td>{mating.sonarRsult}</td>
                                     <td>{mating.expectedDeliveryDate ? mating.expectedDeliveryDate.split('T')[0] : NO_DATE}</td>
-                                    <td onClick={() => editMating(mating._id)} style={{ cursor: 'pointer' }} className="text-primary">
-                                        <FaRegEdit /> Edit Mating 
+                                    
+                                    <td onClick={() => editMating(mating._id)} style={{ cursor: 'pointer' }} className='text-primary'>
+                                        <FaRegEdit /> Edit Mating
                                     </td>
-                                    <td onClick={() => handleClick(mating._id)} className="text-danger" style={{ cursor: 'pointer' }}>
+                                    <td onClick={() => confirmDelete(mating._id)} style={{ cursor: 'pointer' }} className='text-danger'>
                                         <RiDeleteBin6Line /> Remove Mating
                                     </td>
                                 </tr>
@@ -171,38 +183,20 @@ function MatingTable() {
                         </tbody>
                     </table>
                     
-                    {/* Pagination Controls */}
                     <div className="d-flex justify-content-center mt-4">
                         <nav>
                             <ul className="pagination">
-                                {/* Go to first page */}
                                 <li className={`page-item ${currentPage === 1 ? 'disabled' : ''}`}>
                                     <button className="page-link" onClick={() => paginate(1)}>First</button>
                                 </li>
-
-                                {/* Show previous page */}
                                 <li className={`page-item ${currentPage === 1 ? 'disabled' : ''}`}>
                                     <button className="page-link" onClick={() => paginate(currentPage - 1)}>Previous</button>
                                 </li>
-
-                                {/* Dynamic page numbers */}
-                                {Array.from({ length: totalPages }, (_, index) => {
-                                    const pageNumber = index + 1;
-                                    const isPageInRange = pageNumber >= currentPage - 2 && pageNumber <= currentPage + 2;
-
-                                    if (isPageInRange) {
-                                        return (
-                                            <li key={index} className={`page-item ${pageNumber === currentPage ? 'active' : ''}`}>
-                                                <button className="page-link" onClick={() => paginate(pageNumber)}>
-                                                    {pageNumber}
-                                                </button>
-                                            </li>
-                                        );
-                                    }
-                                    return null;
-                                })}
-
-                                {/* Show next page */}
+                                {Array.from({ length: totalPages }, (_, index) => (
+                                    <li key={index} className={`page-item ${currentPage === index + 1 ? 'active' : ''}`}>
+                                        <button className="page-link" onClick={() => paginate(index + 1)}>{index + 1}</button>
+                                    </li>
+                                ))}
                                 <li className={`page-item ${currentPage === totalPages ? 'disabled' : ''}`}>
                                     <button className="page-link" onClick={() => paginate(currentPage + 1)}>Next</button>
                                 </li>
